+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * Fixed issue when "Read more" was overflowing text
 * Fixed issue when logged in user did not see his avatar in editor 
+* Fixed issue when comment count header was not showing after recent update, #199
 
 **Other:**