+++ CHANGELOG.md
@@ -9,8 +9,10 @@
 **Fixes:**
 
 * Fixed issue when "Read more" was overflowing text
-* Fixed issue when logged in user did not see his avatar in editor 
+* Fixed issue when logged in user did not see his avatar in editor
 * Fixed issue when comment count header was not showing after recent update, #199
+* Fixed issue when table for managing notification emails was not created, #196
+* All tables which come from plugin now prefixed as defined by config, #200
 
 **Other:**