+++ CHANGELOG.md
@@ -4,8 +4,6 @@
 
 **Enhancements:**
 
-
-
 **Fixes:**
 
 * Fixed issue when "Read more" was overflowing text
@@ -13,6 +11,8 @@
 * Fixed issue when comment count header was not showing after recent update, #199
 * Fixed issue when table for managing notification emails was not created, #196
 * All tables which come from plugin now prefixed as defined by config, #200
+* Now editor styles also adapt to custom styles defined in "Design" tag (e.g. font family, font size, etc), #195
+* Fixed issue when some themes have special styles for `<p>` or `<ul>` tag which caused comment text to be weirdly formatted, #202
 
 **Other:**