+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.0.60 – 06.10.2018
+
+**Enhancements:**
+
+
+
+**Fixes:**
+
+* Fixed issue when update comment and then reply action still was on the update action when sending comment, #205
+
+
 ## 0.0.59 – 06.10.2018
 
 **Enhancements:**